% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conj_rate.R
\name{.check_cols_present}
\alias{.check_cols_present}
\title{Check if necessary columns are present}
\usage{
.check_cols_present(data, necessary_cols)
}
\arguments{
\item{data}{Dataframe. Required columns vary depending on
the method used to estimate the conjugation rates. The following
columns are typically required: D.t, R.t, T.t are final population sizes;
D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates;
t is the time of measurement (assumes start at 0).
Optional columns: T.0, psi.max (for the Simonsen method).}

\item{necessary_cols}{list of columns to check the presence of.}
}
\description{
Check if necessary columns are present
}
