% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conj_rate.R
\name{.check_conj_input}
\alias{.check_conj_input}
\title{Check conjugation rate estimation input}
\usage{
.check_conj_input(data, method)
}
\arguments{
\item{data}{Dataframe. Required columns vary depending on
the method used to estimate the conjugation rates. The following
columns are typically required: D.t, R.t, T.t are final population sizes;
D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates;
t is the time of measurement (assumes start at 0).
Optional columns: T.0, psi.max (for the Simonsen method).}

\item{method}{String or list of strings. Describes method(s) used to
estimate the conjugation rate. Takes the following
values:
\itemize{
   \item SM (Simonsen Method): The population growth rate can be supplied with
 psi.N or psi.max; otherwise the method will choose
 the maximum among psi.D, psi.R, psi.T;
   \item ASM (Approximate Simonsen Method): ;
   \item TD: \eqn{T/D};
   \item T_DR: \eqn{T/DR};
   \item T_RT: \eqn{T/(R+T)}: The fraction
  of plasmid-carrying recipients;
   \item Dionisio: \eqn{T/\sqrt{DR}};
   \item Gama: \eqn{log10(T/\sqrt{DR})}.
}}
}
\description{
Check conjugation rate estimation input
}
