% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{.check_crit_time_input}
\alias{.check_crit_time_input}
\title{Checks input to estimate critical times}
\usage{
.check_crit_time_input(data, verbose = T)
}
\arguments{
\item{data}{Dataframe combining both DRT and TRT data.}

\item{verbose}{Boolean. Should warnings be returned?}
}
\description{
Checks input to estimate critical times
}
