% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_growth_rate.R
\name{.check_growth_input}
\alias{.check_growth_input}
\title{Check growth input data
\code{.check_growth_input} performs general checks on the input data.}
\usage{
.check_growth_input(OD_data, t_col)
}
\arguments{
\item{OD_data}{Dataframe of OD data. Expects a column for the time (t_col),
and will interpret all other columns as OD data for different samples (identified
by their column name).}

\item{t_col}{Name of the time column (defaults to 'time').}
}
\description{
Check growth input data
\code{.check_growth_input} performs general checks on the input data.
}
