% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conj_rate.R
\name{.estimate_conj_per_method}
\alias{.estimate_conj_per_method}
\title{Subroutine to compute individual conjugation rate estimates}
\usage{
.estimate_conj_per_method(data, method = "ASM", id_cols = c("ID"), verbose = T)
}
\arguments{
\item{data}{Dataframe. Required columns vary depending on
the method used to estimate the conjugation rates. The following
columns are typically required: D.t, R.t, T.t are final population sizes;
D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates;
t is the time of measurement (assumes start at 0).
Optional columns: T.0, psi.max (for the Simonsen method).}

\item{method}{String or list of strings. Describes method(s) used to
estimate the conjugation rate. Takes the following
values:
- SM (Simonsen Method): The population growth rate can be supplied with
 psi.N or psi.max; otherwise the method will choose
 the maximum among psi.D, psi.R, psi.T;
- ASM (Approximate Simonsen Method);
- TD \eqn{T/D};
- T_DR \eqn{T/DR};
- T_RT \eqn{T/(R+T)}: The fraction
  of plasmid-carrying recipients;
- Dionisio;
- Gama.}

\item{id_cols}{List of strings. Column names in the data
that should be treated as identifiers (will be returned in output).}

\item{verbose}{Boolean. Should the method return warnings?}
}
\value{
Dataframe with conjugation rate estimates.
Contains 3 columns: sample IDs, estimates, methods.
}
\description{
\code{.estimate_conj_per_method} returns the conjugation
rate estimates from data (for a single method).
}
