% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conj_rate.R
\name{.estimate_Dionisio}
\alias{.estimate_Dionisio}
\title{Dionisio estimate from data}
\usage{
.estimate_Dionisio(data)
}
\arguments{
\item{data}{Dataframe. Required columns vary depending on
the method used to estimate the conjugation rates. The following
columns are typically required: D.t, R.t, T.t are final population sizes;
D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates;
t is the time of measurement (assumes start at 0).
Optional columns: T.0, psi.max (for the Simonsen method).}
}
\value{
estimate vector
}
\description{
\code{.estimate_Dionisio} returns the
Dionisio estimate: T/sqrt(DR).
}
\details{
Dionisio F, Matic I, Radman M, Rodrigues OR, Taddei F.
Plasmids spread very fast in heterogeneous bacterial communities.
Genetics 2002 Dec;162(4):1525–32.
}
