% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conj_rate.R
\name{.estimate_SM}
\alias{.estimate_SM}
\alias{gamma_max}
\title{Simonsen conjugation rate estimate}
\usage{
.estimate_SM(data)
}
\arguments{
\item{data}{Dataframe. Required columns vary depending on
the method used to estimate the conjugation rates. The following
columns are typically required: D.t, R.t, T.t are final population sizes;
D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates;
t is the time of measurement (assumes start at 0).
Optional columns: T.0, psi.max (for the Simonsen method).}
}
\value{
estimate vector
}
\description{
\code{.estimate_SM} returns the Simonsen
 formula from data.
 The population growth rate can be supplied with
 psi.N or psi.max; otherwise the method will choose
 the maximum among psi.D, psi.R, psi.T.
}
