% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conj_rate.R
\name{.estimate_T_DR}
\alias{.estimate_T_DR}
\title{Estimate T/DR from data}
\usage{
.estimate_T_DR(data)
}
\arguments{
\item{data}{Dataframe. Required columns vary depending on
the method used to estimate the conjugation rates. The following
columns are typically required: D.t, R.t, T.t are final population sizes;
D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates;
t is the time of measurement (assumes start at 0).
Optional columns: T.0, psi.max (for the Simonsen method).}
}
\value{
estimate vector
}
\description{
\code{.estimate_T_DR} returns the T/DR estimate.
}
