% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{.estimate_tcrit1}
\alias{.estimate_tcrit1}
\title{Estimate tcrit1}
\usage{
.estimate_tcrit1(data, tol_factor = 10)
}
\arguments{
\item{data}{Dataframe combining both DRT and TRT data.}

\item{tol_factor}{Double. The imprecision factor: this
factor indicates how close the system is allowed to
approach a state where it violates the ASM assumptions
(smaller values allow greater violation).}
}
\value{
Estimates for tcrit1
}
\description{
Estimate tcrit1
}
