% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{.estimate_tcrit2}
\alias{.estimate_tcrit2}
\title{Estimate tcrit2}
\usage{
.estimate_tcrit2(data, tol_factor = 10)
}
\arguments{
\item{data}{Dataframe combining both DRT and TRT data.}

\item{tol_factor}{Double. The imprecision factor.}
}
\value{
Estimate for tcrit2
}
\description{
Estimate tcrit2
}
