% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{.estimate_tcrit3}
\alias{.estimate_tcrit3}
\title{Estimate tcrit3}
\usage{
.estimate_tcrit3(data, tol_factor = 10)
}
\arguments{
\item{data}{Dataframe combining both DRT and TRT data.}

\item{tol_factor}{Double. The imprecision factor.}
}
\value{
Estimates for tcrit3
}
\description{
Estimate tcrit3
}
