% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{.get_crit_time_input}
\alias{.get_crit_time_input}
\title{Shape data to estimate critical times}
\usage{
.get_crit_time_input(DRT, TRT, id_cols = c("ID"), verbose = T)
}
\arguments{
\item{DRT}{Dataframe. Data corresponding to the first
conjugation experiment.
Necessary columns: D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates.
If gamma.D is not among the columns,
the function will attempt to estimate gamma.D from
the data given. This requires D.t, R.t, T.t
(final population sizes) and t (the time of measurement).
If gamma.T is among the columns, the input under TRT will
be ignored.}

\item{TRT}{Dataframe. Data corresponding to the second
conjugation experiment. Same columns needed as for DRT.}

\item{id_cols}{List of strings. Specifies the column
names that should be treated as identifiers (will be returned in output).}

\item{verbose}{Boolean. Should warnings be returned?}
}
\description{
\code{.get_crit_time_input} returns the dataframe needed
to estimate the critical times corresponding
to a conjugation experiment.
}
\details{
Note: the number of columns in the output can vary,
depending on whether gamma.T can be estimated or not.
}
