% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_growth_rate.R
\name{.logist}
\alias{.logist}
\title{Logistic growth function}
\usage{
.logist(t, carrying_cap, growth_rate, N0)
}
\arguments{
\item{t}{Time vector.}

\item{carrying_cap}{Carrying capacity.}

\item{growth_rate}{Growth rate.}

\item{N0}{Initial population size.}
}
\value{
Final population size after log growth with the given parameters.
}
\description{
\code{.logist} computes the logistic growth function.
}
\seealso{
Other growth rate functions: 
\code{\link{estimate_growth_rate}()},
\code{\link{get_growth_fit_for_plot}()}
}
\concept{growth rate functions}
