% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{estimate_crit_time}
\alias{estimate_crit_time}
\alias{critical_time}
\alias{crit_time}
\title{Estimate critical times}
\usage{
estimate_crit_time(
  DRT,
  TRT = NULL,
  tol_factor = 10,
  id_cols = c("ID"),
  verbose = T
)
}
\arguments{
\item{DRT}{Dataframe. Data corresponding to the first
conjugation experiment.
Necessary columns: D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates.
If gamma.D is not among the columns,
the function will attempt to estimate gamma.D from
the data given. This requires D.t, R.t, T.t
(final population sizes) and t (the time of measurement).
If gamma.T is among the columns, the input under TRT will
be ignored.}

\item{TRT}{Dataframe. Data corresponding to the second
conjugation experiment. Same columns needed as for DRT.}

\item{tol_factor}{Double. The imprecision factor: this
factor indicates how close the system is allowed to
approach a state where it violates the ASM assumptions
(smaller values allow greater violation).}

\item{id_cols}{List of strings. Specifies the column
names that should be treated as identifiers (will be returned in output).}

\item{verbose}{Boolean. Should warnings be returned?}
}
\value{
Dataframe of critical time estimates.
}
\description{
\code{estimate_crit_time} returns the
 conjugation rates and critical times corresponding
 to a conjugation experiment.
 For the ASM to return a proper conjugation estimate
 the time point of measurement must be smaller than
 min(tcrit1, tcrit2, tcrit3).
}
\examples{
estimate_crit_time(DRT_example, TRT_example)
estimate_crit_time(DRT_example, TRT_example, id_cols = c('ID', 't'))
}
\seealso{
Other critical time functions: 
\code{\link{scan_crit_time}()}
}
\concept{critical time functions}
