% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_growth_rate.R
\name{estimate_growth_rate}
\alias{estimate_growth_rate}
\alias{growth}
\alias{growth_rate}
\title{Fit input data to a logistic growth function}
\usage{
estimate_growth_rate(OD_data, t_start = 0, t_col = "time", verbose = T)
}
\arguments{
\item{OD_data}{Dataframe of OD data. Expects a column for the time (t_col),
and will interpret all other columns as OD data for different samples (identified
by their column name).}

\item{t_start}{Earliest possible time point from which to start the fit.}

\item{t_col}{Name of the time column (defaults to 'time').}

\item{verbose}{Boolean. Should warnings be returned?}
}
\value{
Dataframe of estimated growth parameters. Contains columns
name, carrying_cap, growth_rate, N0, t_min and t_stat.
}
\description{
\code{estimate_growth_rate} fits the input OD data to a
  logistic growth function.
It differs from growthcurver by fitting the logged data and function.
It also subtracts an estimated value for blank wells (taking the minimum
across all readings), and takes into account measurements only
after t_start.
}
\examples{
estimate_growth_rate(growth_example, t_start = 1, t_col = 'Time_h')
}
\seealso{
Other growth rate functions: 
\code{\link{.logist}()},
\code{\link{get_growth_fit_for_plot}()}
}
\concept{growth rate functions}
