% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_growth_rate.R
\name{get_growth_fit_for_plot}
\alias{get_growth_fit_for_plot}
\alias{growth_fit}
\title{Growth fit for plotting
\code{get_growth_fit_for_plot} computes a vector of
the fit of the growth data, for comparison against the
OD data.}
\usage{
get_growth_fit_for_plot(OD_data, t_col = "time", growth_fit, verbose = T)
}
\arguments{
\item{OD_data}{Dataframe of OD data. Expects a column for the time (t_col),
and will interpret all other columns as OD data for different samples (identified
by their column name).}

\item{t_col}{Name of the time column (defaults to 'time').}

\item{growth_fit}{Output of estimate_growth_rate for the OD_data;
requires columns name, carrying_cap, growth_rate, N0, t_min, t_stat}

\item{verbose}{Boolean. Should warnings be returned?}
}
\value{
Simulated log growth matching the input fits.
}
\description{
Growth fit for plotting
\code{get_growth_fit_for_plot} computes a vector of
the fit of the growth data, for comparison against the
OD data.
}
\examples{
growth_fit <- estimate_growth_rate(growth_example, t_start = 1, t_col = 'Time_h')
get_growth_fit_for_plot(growth_example, t_col = 'Time_h', growth_fit)
}
\seealso{
Other growth rate functions: 
\code{\link{.logist}()},
\code{\link{estimate_growth_rate}()}
}
\concept{growth rate functions}
