% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_example}
\alias{growth_example}
\title{Example OD measurements}
\format{
A data frame with 3 rows and 10 variables:
\describe{
  \item{Time_h}{time in hours}
  \item{Donor_OD}{OD of the donors, over 12 hours}
  \item{Recipient_OD}{OD of the recipients, over 12 hours}
}
}
\usage{
growth_example
}
\description{
Example OD measurements
}
\keyword{datasets}
