% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_crit_time.R
\name{scan_crit_time}
\alias{scan_crit_time}
\alias{scan_crit}
\alias{time_scan}
\alias{crittime_scan}
\title{Scan critical times}
\usage{
scan_crit_time(
  DRT,
  tol_factor = 10,
  id_cols = c("ID"),
  mult_seq = 10^seq(-2, 5, 1),
  verbose = T
)
}
\arguments{
\item{DRT}{Dataframe. Data corresponding to the first
conjugation experiment.
Necessary columns: D.0, R.0 are initial pop sizes;
psi.D, psi.R, psi.T are growth rates.
If gamma.D is not among the columns,
the function will attempt to estimate gamma.D from
the data given. This requires D.t, R.t, T.t
(final population sizes) and t (the time of measurement).
If gamma.T is among the columns, the input under TRT will
be ignored.}

\item{tol_factor}{Double. The imprecision factor: this
factor indicates how close the system is allowed to
approach a state where it violates the ASM assumptions
(smaller values allow greater violation).}

\item{id_cols}{List of strings. Specifies the column
names that should be treated as identifiers (will be returned in output).}

\item{mult_seq}{Vector. Describes the multiplication factors
to take into account for gamma.T.}

\item{verbose}{Boolean. Should warnings be returned?}
}
\value{
Dataframe of estimated critical times for the input DRT
data and a range of gamma.T values.
}
\description{
\code{scan_crit_time} returns the
conjugation rates and critical times corresponding to
a DRT conjugation experiment, assuming different gamma.T values.
 For the ASM to return a proper conjugation estimate
 the time point of measurement must be smaller than
 min(tcrit1, tcrit2, tcrit3).
}
\examples{
scan_crit_time(DRT_example)
scan_crit_time(DRT_example, id_cols = c('ID', 't'), mult_seq = 10**seq(0, 3, 0.1))
}
\seealso{
\code{\link{estimate_conj_rate}} for estimation of the
conjugation rates.

Other critical time functions: 
\code{\link{estimate_crit_time}()}
}
\concept{critical time functions}
