% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRT_example}
\alias{TRT_example}
\title{Example conjugation rates (TRT)}
\format{
A data frame with 3 rows and 10 variables:
\describe{
  \item{ID}{sample ID}
  \item{psi.D}{growth rate of the donors, per hour}
  \item{psi.R}{growth rate of the recipients, per hour}
  \item{psi.T}{growth rate of the transconjugants, per hour}
  \item{D.0}{initial population size of the donors, in CFU/mL}
  \item{R.0}{initial population size of the recipients, in CFU/mL}
  \item{D.t}{final population size of the donors, in CFU/mL}
  \item{R.t}{final population size of the recipients, in CFU/mL}
  \item{T.t}{final population size of the transconjugants, in CFU/mL}
  \item{t}{Measurement timepoint, in hours}
}
}
\usage{
TRT_example
}
\description{
An example TRT dataset containing made up conjugation rates
}
\keyword{datasets}
